Debuging

+++ lib/cogs/mod.py
@@ -1,15 +1,16 @@
+import re
 from typing import Optional
 from datetime import datetime
 from datetime import timedelta
-from re import search
 from asyncio import sleep
-
+from re import search
 
 from better_profanity import profanity
 import discord
 from discord.ext.commands import command
 from discord.ext.commands import Cog, Greedy
 from discord.ext.commands import has_permissions, bot_has_permissions, CheckFailure
+from discord import Forbidden
 from discord import Embed, Member
 
 from ..db import db
@@ -21,29 +22,30 @@ class Mod(Cog):
         self.bot = bot
 
         self.url_regex = r"(?i)\b((?:https?://|www\d{0,3}[.]|[a-z0-9.\-]+[.][a-z]{2,4}/)(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:'\".,<>?«»“”‘’]))"
-        self.no_links = (817805249552187443, 688498139459879115, 736932762615021660, 736934235461517403, 737693038029307975, 707691743986057287) 
+        self.no_links = (817805249552187443, 688498139459879115, 736932762615021660, 736934235461517403, 737693038029307975, 707691743986057287)
             
-    async def kick_members(self, message, targets, reason):
-        for target in targets:
-            if (message.guild.me.top_role.position > target.top_role.position
-                and not target.guild_permissions.administrator):
-                await target.kick(reason=reason)
+    async def kick_members(self, ctx, targets, reason):
+        if not len(targets):
+            await ctx.send("One or more required arguments are missing.")
+        else:
+            for target in targets:
+                if ctx.guild.me.top_role.position > target.top_role.position and not target.guild_permissions.administrator:
+                    try:
+                        await target.send(f"You were kicked by {ctx.author.display_name} because {reason}")
+                    except Forbidden:
+                        pass
+                    await target.kick(reason=reason)
+                    embed=Embed(title="Member kicked", colour=0xff0000, timestamp=datetime.utcnow())
+                    embed.set_author(name=ctx.author.display_name)
+                    embed.set_thumbnail(url=target.avatar_url)
+                    embed.add_field(name="Member", value=f"{target.name} a.k.a. {target.display_name}", inline=False),
+                    embed.add_field(name="Actioned by", value=ctx.author.display_name, inline=False),
+                    embed.add_field(name="Reason", value=reason, inline=False)
+                    await self.logs_channel.send(embed=embed)
 
-                embed = Embed(title="Member kicked",
-                              colour=0xDD2222,
-                              timestamp=datetime.utcnow())
-
-                embed.set_thumbnail(url=target.avatar_url)
-
-                fields = [("Member", f"{target.name} a.k.a. {target.display_name}", False),
-                          ("Actioned by", message.author.display_name, False),
-                          ("Reason", reason, False)]
-
-                for name, value, inline in fields:
-                    embed.add_field(name=name, value=value, inline=inline)
-
-                await self.logs_channel.send(embed=embed)
-
+                else:
+                    await ctx.send(f"Somehow {target.display_name} could not be kicked.")
+    
     @command(name="kick")
     @bot_has_permissions(kick_members=True)
     @has_permissions(kick_members=True)
@@ -54,33 +56,36 @@ class Mod(Cog):
         else:
             await self.kick_members(ctx.message, targets, reason)
             await ctx.send("Action complete.")
-
+    
     @kick_command.error
     async def kick_command_error(self, ctx, exc):
         if isinstance(exc, CheckFailure):
-            await ctx.send("Insufficient permissions to perform that task.")
+            await ctx.send("You don't have permission for that.")
 
-    async def ban_members(self, message, targets, reason):
-        for target in targets:
-            if (message.guild.me.top_role.position > target.top_role.position
-                and not target.guild_permissions.administrator):
-                await target.ban(reason=reason)
+###############################################################################################
 
-                embed = Embed(title="Member banned",
-                              colour=0xDD2222,
-                              timestamp=datetime.utcnow())
+    async def ban_members(self, ctx, targets, reason):
+        if not len(targets):
+            await ctx.send("One or more required arguments are missing.")
+        else:
+            for target in targets:
+                if ctx.guild.me.top_role.position > target.top_role.position and not target.guild_permissions.administrator:
+                    try:
+                        await target.send(f"You were banned by {ctx.author.display_name} because {reason}")
+                    except Forbidden:
+                        pass
+                    await target.ban(reason=reason)
+                    embed=Embed(title="Member banned", colour=0xff0000, timestamp=datetime.utcnow())
+                    embed.set_author(name=ctx.author.display_name)
+                    embed.set_thumbnail(url=target.avatar_url)
+                    embed.add_field(name="Member", value=f"{target.name} a.k.a. {target.display_name}", inline=False),
+                    embed.add_field(name="Actioned by", value=ctx.author.display_name, inline=False),
+                    embed.add_field(name="Reason", value=reason, inline=False)
+                    await self.logs_channel.send(embed=embed)
 
-                embed.set_thumbnail(url=target.avatar_url)
-
-                fields = [("Member", f"{target.name} a.k.a. {target.display_name}", False),
-                          ("Actioned by", message.author.display_name, False),
-                          ("Reason", reason, False)]
-
-                for name, value, inline in fields:
-                    embed.add_field(name=name, value=value, inline=inline)
-
-                await self.logs_channel.send(embed=embed)
-
+                else:
+                    await ctx.send(f"Somehow {target.display_name} could not be banned.")
+    
     @command(name="ban")
     @bot_has_permissions(ban_members=True)
     @has_permissions(ban_members=True)
@@ -91,13 +96,15 @@ class Mod(Cog):
         else:
             await self.ban_members(ctx.message, targets, reason)
             await ctx.send("Action complete.")
-
+    
     @ban_command.error
     async def ban_command_error(self, ctx, exc):
         if isinstance(exc, CheckFailure):
-            await ctx.send("Insufficient permissions to perform that task.")
+            await ctx.send("You don't have permission for that.")
 
-    @command(name="clear", aliases=["purge"])
+##############################################################################################################################################
+
+    @command(name="clear", aliases=["purge", "nuke"])
     @bot_has_permissions(manage_messages=True)
     @has_permissions(manage_messages=True)
     async def clear_messages(self, ctx, targets: Greedy[Member], limit: Optional[int] = 1):
@@ -113,63 +120,59 @@ class Mod(Cog):
                 await ctx.send(f"Deleted {len(deleted):,} messages.", delete_after=5)
 
         else:
-            await ctx.send("The limit provided is not within acceptable bounds.")
+            await ctx.send("The limit provided isn't within acceptable bounds. You can clear from 1 to 100.")
 
+##############################################################################################################################################
 
-    # def days_hours_minutes(td):
-    #     return td.days, td.seconds//3600, (td.seconds//60)%60
-
-    async def mute_members(self, message, targets, seconds, reason):
+    async def mute_members(self, message, targets, hours, reason):
         unmutes = []
-        minutes = seconds*60
 
         for target in targets:
             if not self.mute_role in target.roles:
                 if message.guild.me.top_role.position > target.top_role.position:
                     role_ids = ",".join([str(r.id) for r in target.roles])
-                    end_time = datetime.utcnow() + timedelta(minutes) if minutes else None
+                    end_time = datetime.utcnow() + timedelta(seconds=hours) if hours else None
 
                     db.execute("INSERT INTO mutes VALUES (?, ?, ?)",
-                               target.id, role_ids, getattr(end_time, "isoformat", lambda: None)())
+                            target.id, role_ids, getattr(end_time, "isoformat", lambda: None)())
 
                     await target.edit(roles=[self.mute_role])
 
                     embed = Embed(title="Member muted",
-                                  colour=0xDD2222,
-                                  timestamp=datetime.utcnow())
+                                colour=0xDD2222,
+                                timestamp=datetime.utcnow())
 
                     embed.set_thumbnail(url=target.avatar_url)
 
                     fields = [("Member", target.display_name, False),
-                              ("Actioned by", message.author.display_name, False),
-                              ("Duration", f"{minutes:,} minute(s)" if minutes else "Indefinite", False),
-                              ("Reason", reason, False)]
+                            ("Actioned by", message.author.display_name, False),
+                            ("Duration", f"{hours:,} hour(s)" if hours else "Indefinite", False),
+                            ("Reason", reason, False)]
 
                     for name, value, inline in fields:
                         embed.add_field(name=name, value=value, inline=inline)
 
                     await self.logs_channel.send(embed=embed)
 
-                    if minutes:
+                    if hours:
                         unmutes.append(target)
 
         return unmutes
 
     @command(name="mute")
     @bot_has_permissions(manage_roles=True)
-    @has_permissions(manage_roles=True, manage_guild=True)
-    async def mute_command(self, ctx, targets: Greedy[Member], minutes: Optional[int], *,
+    @has_permissions(manage_messages=True)
+    async def mute_command(self, ctx, targets: Greedy[Member], hours: Optional[int], *,
                            reason: Optional[str] = "No reason provided."):
         if not len(targets):
             await ctx.send("One or more required arguments are missing.")
 
         else:
-            unmutes = await self.mute_members(ctx.message, targets, minutes, reason)
+            unmutes = await self.mute_members(ctx.message, targets, hours, reason)
             await ctx.send("Action complete.")
 
             if len(unmutes):
-                # minutes = minutes * 60
-                await sleep(minutes)
+                await sleep(hours)
                 await self.unmute_members(ctx.guild, targets)
 
     @mute_command.error
@@ -203,7 +206,7 @@ class Mod(Cog):
 
     @command(name="unmute")
     @bot_has_permissions(manage_roles=True)
-    @has_permissions(manage_roles=True, manage_guild=True)
+    @has_permissions(manage_messages=True)
     async def unmute_command(self, ctx, targets: Greedy[Member], *, reason: Optional[str] = "No reason provided."):
         if not len(targets):
             await ctx.send("One or more required arguments is missing.")
@@ -236,31 +239,20 @@ class Mod(Cog):
         embed=discord.Embed(title="Profanity deleted", description="Word was deleted from the list", color=0xc91313)
         await ctx.send(embed=embed)
 
+
     @Cog.listener()
     async def on_message(self, message):
-        def _check(m):
-            return (m.author == message.author
-                    and len(m.mentions)
-                    and (datetime.utcnow()-m.created_at).seconds < 60)
-
         if not message.author.bot:
-            if len(list(filter(lambda m: _check(m), self.bot.cached_messages))) >= 3:
-                await message.channel.send("Don't spam mentions!", delete_after=10)
-                unmutes = await self.mute_members(message, [message.author], 5, reason="Mention spam")
-                print("Mention spam")
+            if not message.author.guild_permissions.manage_guild:
+                if profanity.contains_profanity(message.content):
+                    await message.delete()
+                    await message.channel.send("You can't use that word here!", delete_after=10)
 
-                if len(unmutes):
-                    await sleep(5)
-                    await self.unmute_members(message.guild, [message.author])
+                elif message.channel.id not in self.links_allowed and search(self.url_regex, message.content):
+                    await message.delete()
+                    await message.channel.send("You can't send links in this channel.", delete_after=10)
 
-            elif profanity.contains_profanity(message.content):
-                await message.delete()
-                await message.channel.send("You can't use that word here.", delete_after=10)
 
-            elif message.channel.id in self.no_links and search(self.url_regex, message.content):
-                await message.delete()
-                await message.channel.send("You can't send links in this channel.", delete_after=10)
-    
     @Cog.listener()
     async def on_ready(self):
         if not self.bot.ready:
